chrome_dev_tool_remote/pyproject.toml
[project]
name = "chrome-dev-tool-remote"
version = "0.1.0"
description = "Chrome DevTools remote log collector with marimo UI"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.21.1",
    "duckdb>=1.0.0",
    "pandas>=2.2.0",
    "websockets>=13.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.18.0",
    "altair>=5.2.0",
    "paramiko>=4.0.0",
]
